+++ apps/doc/src/app/components/table/table-example.module.ts
@@ -35,9 +35,11 @@ import { TableLoadingExampleComponent } from './examples/table-loading-example/t
 import { TableStickyExampleComponent } from './examples/table-sticky-example/table-sticky-example.component';
 import { TableDataSourceExampleComponent } from './examples/table-data-source-example/table-data-source-example.component';
 import { TableDynamicDynamicRowGroupExampleComponent } from './examples/table-dynamic-row-group-example/table-dynamic-row-group-example.component';
+import { TableInheritColsExampleComponent } from './examples/table-inherit-cols/table-inherit-cols-example.component';
 
 @NgModule({
   declarations: [
+    TableInheritColsExampleComponent,
     TableExampleComponent,
     TableBasicExampleComponent,
     TableLoadingExampleComponent,