+++ apps/doc/src/app/about-prizm/changelog/CHANGELOG.md
@@ -8,6 +8,10 @@ All notable changes to this project will be documented in this file.
 
 - feat(components/table): add support dynamic amount of columns in group #259
 
+### Bug fixes
+
+- fix(components/table): order of [columns] not inherited from table #359
+
 ## [2.1.4-next.1](https://github.com/zyfra/Prizm) (02-08-2023)
 
 ### Features